server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
Pillow==10.2.0
openpyxl==3.1.2
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.26.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
